Add cross-correlation analysis scripts
plotInstantaneousDelays shows a 'snapshot' of the position and heading
of the focal participant and each of his/her neighbors; time delays are
plotted next to each neighbor, and red lines are drawn to 'leading'
neighbors. plotMeanDelayHeatmap essentially replicates this analysis
several times, and generates a heatmap showing the mean time delay as a
spatial distribution centered around the focal participant.